Information for each electrofishing sample from Big Hill Reservoir, KS, 2014.

Information for each electrofishing sample from Big Hill Reservoir, KS, in May, 2014.

Format

A data frame with 20 observations on the following 4 variables.

UID

Unique sample identification number

date

Data sample was collected

loc

Location code for where the sample was collected

effort

Effort (minutes) expended for the sample

Source

Obtained directly from Ben Neely. CSV file

Note

Used in the Introductory Fisheries Analyses with R book.

Topic(s)

Data Manipulation

See also

Examples

data(BGHRsample)
str(BGHRsample)
#> 'data.frame':	20 obs. of  4 variables:
#>  $ UID   : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ date  : Factor w/ 2 levels "5/14/2014","5/15/2014": 2 2 1 1 2 1 2 2 1 2 ...
#>  $ loc   : int  196 200 201 207 219 220 225 227 228 229 ...
#>  $ effort: int  10 10 10 10 10 10 10 10 10 10 ...
head(BGHRsample)
#>   UID      date loc effort
#> 1   1 5/15/2014 196     10
#> 2   2 5/15/2014 200     10
#> 3   3 5/14/2014 201     10
#> 4   4 5/14/2014 207     10
#> 5   5 5/15/2014 219     10
#> 6   6 5/14/2014 220     10