Catch and effort data for male Blue Crabs.

Catch and effort data for a population of male Blue Crabs (Callinectes sapidus) for a 12-week period.

Format

A data frame with 12 observations on the following 2 variables.

catch

A numeric vector of pounds of Blue Crab caught.

effort

A numeric vector of lines used per day to catch Blue Crab.

Source

From Fischler, K.J. 1965. The use of catch-effort, catch-sampling, and tagging data to estimate a population of blue crabs. Transactions of the American Fisheries Society 94:287-310. CSV file

Topic(s)

Population size

Abundance

Depletion

Leslie method

DeLury method

Catchability

Examples

data(BlueCrab)
str(BlueCrab)
#> 'data.frame':	12 obs. of  2 variables:
#>  $ catch : int  33541 47326 36460 33157 29207 33125 14191 9503 13115 13663 ...
#>  $ effort: int  194 248 243 301 357 352 269 244 256 258 ...
BlueCrab
#>    catch effort
#> 1  33541    194
#> 2  47326    248
#> 3  36460    243
#> 4  33157    301
#> 5  29207    357
#> 6  33125    352
#> 7  14191    269
#> 8   9503    244
#> 9  13115    256
#> 10 13663    258
#> 11 10865    234
#> 12  9887    227