Capture histories (9 samples) of Cutthroat Trout from Auke Lake.

Summarized (“RMark” format) capture histories of Cutthroat Trout (Oncorhynchus clarki) in Auke Lake, Alaska, from samples taken in 1998-2006.

Format

A data frame with 47 observations on the following 2 variables.

ch

Unique capture history (as a character string)

freq

Frequency of fish with that capture history

Source

Entered from Appendix A.3 of Harding, R.D., C.L. Hoover, and R.P. Marshall. 2010. Abundance of Cutthroat Trout in Auke Lake, Southeast Alaska, in 2005 and 2006. Alaska Department of Fish and Game Fisheries Data Series No. 10-82. Accessed from http://www.sf.adfg.state.ak.us/FedAidPDFs/FDS10-82.pdf. CSV file

Topic(s)

Population Size

Abundance

Mark-Recapture

Capture-Recapture

Jolly-Seber

Capture History

See also

Examples

data(CutthroatALf)
str(CutthroatALf)
#> 'data.frame':	46 obs. of  2 variables:
#>  $ ch  : chr  "000000001" "000000010" "000000011" "000000100" ...
#>  $ freq: num  122 73 9 168 2 28 1 193 1 6 ...
head(CutthroatALf)
#>          ch freq
#> 1 000000001  122
#> 2 000000010   73
#> 3 000000011    9
#> 4 000000100  168
#> 5 000000101    2
#> 6 000000110   28