Catches in removal events of Cutthroat Trout and Coho Salmon in Little Stawamus Creek (British Columbia, Canada) in 1997.

Catches of Coho Salmon (Oncorhynchus kisutch) and Cutthroat Trout (Oncorhynchus clarki) in consecutive removal events at various locations in Little Stawamus Creek (British Columbia, Canada) in 1997.

Format

A data frame of 26 observations on the following 10 variables:

reach

Reach number of sampling location.

habitat

Habitat type of sampling location -- pool, riffle, or run.

strata

Stratum number of sampling location.

area

Area (m^2) of sampling location.

coho1

Coho Salmon removed on the first pass.

coho2

Coho Salmon removed on the second pass.

coho3

Coho Salmon removed on the third pass.

cutt1

Cutthroat Trout removed on the first pass.

cutt2

Cutthroat Trout removed on the second pass.

cutt3

Cutthroat Trout removed on the third pass.

Source

From Appendix 2a and 2b in Decker, A.S., J.M. Bratty, S.C. Riley, and J. Korman. 1999. Estimating standing stock of juvenile coho salmon (Oncorhynchus kisutch) and cutthroat trout (Oncorhynchus clarki) in a small stream: a comparison of sampling designs. Canadian Technical Report of Fisheries and Aquatic Sciences 2282. 24 pp. CSV file

Topic(s)

Population size

Abundance

Removal

Examples

data(Deckeretal1999)
str(Deckeretal1999)
#> 'data.frame':	36 obs. of  10 variables:
#>  $ reach  : int  1 1 1 1 1 1 1 2 2 2 ...
#>  $ habitat: Factor w/ 3 levels "pool","riffle",..: 1 1 2 2 3 3 3 1 1 1 ...
#>  $ strata : int  2 2 1 1 2 2 2 5 5 5 ...
#>  $ area   : int  89 107 51 223 116 46 139 82 103 71 ...
#>  $ coho1  : int  96 62 15 71 85 29 129 377 253 245 ...
#>  $ coho2  : int  23 17 7 17 24 14 23 80 90 39 ...
#>  $ coho3  : int  21 8 0 8 22 7 13 32 48 30 ...
#>  $ cutt1  : int  12 23 4 27 4 13 20 55 61 38 ...
#>  $ cutt2  : int  4 5 2 8 3 2 8 14 20 18 ...
#>  $ cutt3  : int  0 2 0 5 1 0 3 13 15 8 ...
head(Deckeretal1999)
#>   reach habitat strata area coho1 coho2 coho3 cutt1 cutt2 cutt3
#> 1     1    pool      2   89    96    23    21    12     4     0
#> 2     1    pool      2  107    62    17     8    23     5     2
#> 3     1  riffle      1   51    15     7     0     4     2     0
#> 4     1  riffle      1  223    71    17     8    27     8     5
#> 5     1     run      2  116    85    24    22     4     3     1
#> 6     1     run      2   46    29    14     7    13     2     0

## extract data for one sampling location (e.g., 3rd row)
Deckeretal1999[3,]
#>   reach habitat strata area coho1 coho2 coho3 cutt1 cutt2 cutt3
#> 3     1  riffle      1   51    15     7     0     4     2     0