Catches in removal events of Coho Salmon and Dolly Varden Char at various locations near the Greens Creek (AK) Mine site.

Catches in removal events of Coho Salmon (Oncorhynchus kisutch) and Dolly Varden Char (Salvelinus malma) at various locations near the Greens Creek (AK) Mine site.

Format

A data frame of 66 observations on the following 8 variables:

location

Sampling location.

year

Sampling year.

species

Species (Coho.Salmon or Dolly.Varden).

set1

Catch on the first removal pass.

set2

Catch on the second removal pass.

set3

Catch on the third removal pass.

min.FL

Minimum observed fork length.

max.FL

Maximum observed fork length.

Source

From Appendix C1 of Kanouse, K.M. and B.P. Brewster. 2012. Aquatic Biomonitoring at Greens Creek Mine, 2012. Alaska Department of Fish and Game Technical Report No. 12-11. CSV file

Details

Reaches were isolated by natural features, such as shallow riffles. The sample reaches were saturated with 6.35 mm (0.25 in) minnow traps baited with whirl packs containing disinfected salmon eggs.The traps were deployed for 1.5 h and then retrieved where each fish was transferred into a plastic bucket, and the trap was re-baited ans re-set for another 1.5 h soak. In between trapping events, fish were processed -- measured and recorded FL to the nearest 1 mm, weight to the nearest 0.1 g, and species identified. Captured fish were retained during the sample period and returned alive after all three passes were complete.

Topic(s)

Population size

Abundance

Removal

Examples

data(GreensCreekMine)
str(GreensCreekMine)
#> 'data.frame':	66 obs. of  8 variables:
#>  $ location: Factor w/ 4 levels "Greens48","Greens54",..: 1 3 3 2 2 4 4 1 2 2 ...
#>  $ year    : int  2001 2001 2001 2001 2001 2001 2001 2002 2002 2002 ...
#>  $ species : Factor w/ 2 levels "Coho.Salmon",..: 2 2 1 2 1 2 1 2 2 1 ...
#>  $ set1    : int  30 80 1 70 2 70 89 74 168 14 ...
#>  $ set2    : int  16 8 0 49 6 4 18 29 72 6 ...
#>  $ set3    : int  22 43 2 19 4 7 11 23 31 1 ...
#>  $ min.FL  : int  48 52 81 27 32 58 39 45 33 59 ...
#>  $ max.FL  : int  139 168 90 162 95 110 101 160 160 85 ...
head(GreensCreekMine)
#>     location year      species set1 set2 set3 min.FL max.FL
#> 1   Greens48 2001 Dolly.Varden   30   16   22     48    139
#> 2    Greens6 2001 Dolly.Varden   80    8   43     52    168
#> 3    Greens6 2001  Coho.Salmon    1    0    2     81     90
#> 4   Greens54 2001 Dolly.Varden   70   49   19     27    162
#> 5   Greens54 2001  Coho.Salmon    2    6    4     32     95
#> 6 Tributary9 2001 Dolly.Varden   70    4    7     58    110

## extract data for one location, year, and species (e.g., 3rd row)
GreensCreekMine[3,]
#>   location year     species set1 set2 set3 min.FL max.FL
#> 3  Greens6 2001 Coho.Salmon    1    0    2     81     90