Lengths and weights for fish captured in Inch Lake

Total lengths and weights for a subsample of fish captured in Inch Lake, WI in May, 2007 and May, 2008.

Format

A data frame of 516 observations on the following 6 variables:

netID

A unique identifier for the sampling event

fishID

A unique identifier for the individual fish

species

Species name

length

Total length (inches to nearest 0.1)

weight

Wet weight (grams to nearest 0.1)

year

Year of capture

Source

Derek H. Ogle, personal collection

Topic(s)

Weight-Length

Condition

Length Frequency

See also

Examples

data(InchLake2)
str(InchLake2)
#> 'data.frame':	516 obs. of  6 variables:
#>  $ netID  : int  206 205 205 205 205 205 205 205 205 205 ...
#>  $ fishID : int  501 502 503 504 505 506 507 508 509 510 ...
#>  $ species: Factor w/ 9 levels "Black Crappie",..: 2 2 2 2 2 2 2 2 2 2 ...
#>  $ length : num  1.5 1.7 2.2 2.1 1.5 1.9 1.9 1.4 1.2 1.4 ...
#>  $ weight : num  0.7 1.4 1.5 1.4 1 1.8 1.4 0.6 0.3 0.8 ...
#>  $ year   : int  2008 2008 2008 2008 2008 2008 2008 2008 2008 2008 ...
head(InchLake2)
#>   netID fishID  species length weight year
#> 1   206    501 Bluegill    1.5    0.7 2008
#> 2   205    502 Bluegill    1.7    1.4 2008
#> 3   205    503 Bluegill    2.2    1.5 2008
#> 4   205    504 Bluegill    2.1    1.4 2008
#> 5   205    505 Bluegill    1.5    1.0 2008
#> 6   205    506 Bluegill    1.9    1.8 2008

## Isolate just Bluegills
bg.il <- subset(InchLake2,species=="Bluegill")

## Isolate just largemouth bass from 2007
lmb7.il <- subset(InchLake2,species=="Largemouth Bass" & year==2007)