Catches in removal events of Brown and Rainbow Trout at various locations.

Catches of Brown (Salmo trutta) and Rainbow Trout (Oncorhynchus mykiss) in consecutive removal events at various locations.

Format

A data frame of 40 observations on the following 10 variables:

species

Species of trout (brown or rainbow).

site

Site in the watershed. See source.

age0

Logical is TRUE if age-0 and FALSE if age is >0.

first

Catch on the first removal pass.

second

Catch on the second removal pass.

third

Catch on the third removal pass.

pop.cs

Population estimate by Carle-Strub method.

pop.sch

Population estimate by Schnute method.

q.cons

Logical is TRUE if catchability was constant.

rejected

Logical is TRUE if Schnute method rejected the population estimate because the standard error was too large.

Source

From Table 1 in Jones, M.L. and J.D. Stockwell. 1995. A rapid assessment procedure for enumeration of salmonine populations in streams. North American Journal of Fisheries Management, 15:551-562. CSV file

Topic(s)

Population size

Abundance

Removal

Examples

data(JonesStockwell)
str(JonesStockwell)
#> 'data.frame':	40 obs. of  10 variables:
#>  $ species : Factor w/ 2 levels "brown","rainbow": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ site    : Factor w/ 12 levels "A10","A15","A9",..: 2 3 1 4 4 6 6 7 7 8 ...
#>  $ age0    : logi  FALSE TRUE TRUE TRUE FALSE TRUE ...
#>  $ first   : int  17 17 25 69 6 38 13 250 17 71 ...
#>  $ second  : int  10 12 10 43 2 20 10 166 5 48 ...
#>  $ third   : int  7 8 8 23 1 20 4 142 6 40 ...
#>  $ pop.cs  : int  40 47 48 164 9 108 30 927 30 250 ...
#>  $ pop.sch : num  42.1 50 48.5 166 9 114 31.5 942 31 260 ...
#>  $ q.cons  : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ rejected: logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
head(JonesStockwell)
#>   species site  age0 first second third pop.cs pop.sch q.cons rejected
#> 1 rainbow  A15 FALSE    17     10     7     40    42.1   TRUE    FALSE
#> 2 rainbow   A9  TRUE    17     12     8     47    50.0   TRUE    FALSE
#> 3 rainbow  A10  TRUE    25     10     8     48    48.5   TRUE    FALSE
#> 4 rainbow   B1  TRUE    69     43    23    164   166.0   TRUE    FALSE
#> 5 rainbow   B1 FALSE     6      2     1      9     9.0   TRUE    FALSE
#> 6 rainbow Bio1  TRUE    38     20    20    108   114.0   TRUE    FALSE

## extract data for one species, age, and site (e.g., 3rd row)
JonesStockwell[3,]
#>   species site age0 first second third pop.cs pop.sch q.cons rejected
#> 3 rainbow  A10 TRUE    25     10     8     48    48.5   TRUE    FALSE