Catch and effort data for Hawaiian Islands Slipper Lobster.

Catches of Slipper Lobster (Scyllarides squammosus) in three categories from the vicinity of Laysan Bank, Hawaiian Islands on 34 consecutive days in 1986.

Format

A data frame with 34 observations on the following 6 variables.

day

Day of the catch

legal

Number of legal lobsters caught.

sublegal

Number of sub-legal lobsters caught.

berried

Number of egg-bearing lobsters caught.

total

Total number of lobsters caught.

effort

Total daily effort expended.

Source

From Table 1 of Clarke, R.P., and S.S. Yoshimoto. 1990. Application of the Leslie model to commercial catch and effort of the slipper lobster, Scyllarides squammosus, fishery in the northwestern hawaiian islands. Marine Fisheries Review, 52(2):1-7.

Details

Catch (numbers) of lobster in three categories - legal (tail weight greater than 85g), sublegal (tail weight less than 85g), and berried (egg-bearing females). Sublegal and berried lobsters were returned to the water.

The vessel fished between 11 June and 14 July 1986 in the vicinity of Laysan Island and its associated bank. The daily operations of the vessel involved deploying and hauling 1,125 Fathom Plus lobster traps set in strings spaced at 30 m intervals. They were fished in 7 strings of about 160 traps each and baited with Pacific Mackerel, Scomber japonicus. Strings were soaked overnight and retrieved the following day; therefore, the standard unit of effort is the trap-haul.

Topic(s)

Population size

Abundance

Depletion methods

Leslie method

DeLury method

Catchability

Examples

data(LobsterHI)
str(LobsterHI)
#> 'data.frame':	34 obs. of  6 variables:
#>  $ day     : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ legal   : int  2133 1675 2979 3788 2883 2430 4092 3267 3440 2790 ...
#>  $ sublegal: int  1000 450 1785 1850 1750 1470 2400 1800 1000 820 ...
#>  $ berried : int  1000 600 1050 1580 1150 980 1600 1200 1500 450 ...
#>  $ total   : int  4133 2725 5811 7218 5783 4880 8092 6267 5940 4060 ...
#>  $ effort  : int  1125 1125 1125 1125 1125 1125 1125 1125 1125 1125 ...
head(LobsterHI)
#>   day legal sublegal berried total effort
#> 1   1  2133     1000    1000  4133   1125
#> 2   2  1675      450     600  2725   1125
#> 3   3  2979     1785    1050  5811   1125
#> 4   4  3788     1850    1580  7218   1125
#> 5   5  2883     1750    1150  5783   1125
#> 6   6  2430     1470     980  4880   1125