Summarized mark-recapture data for Muskellunge from many Wisconsin Lakes, 2006.

The number of Muskellunge (Esox masquinongy) from a variety of lakes in Wisconsin, 2006.

Format

A data frame with 40 observations on the following 7 variables.

county

County of the collection

lake

Lake of the collection

sex

Sex of the Muskellunge (male, female, unknown, total)

mark

Total number of fish marked on the marking run

catch

Total number of fish caught on the recapture run

recap

Total number of previously marked fish in the recapture run

Source

Wisconsin Department of Natural Resources.

Topic(s)

Population Size

Abundance

Mark-Recapture

Capture-Recapture

Petersen

Examples

data(MuskieWI06MR)
str(MuskieWI06MR)
#> 'data.frame':	40 obs. of  6 variables:
#>  $ county: Factor w/ 5 levels "Langlade","Oneida",..: 1 1 1 1 2 2 2 2 2 2 ...
#>  $ lake  : Factor w/ 10 levels "Big.Arbor.Vitae",..: 4 4 4 4 3 3 3 3 8 8 ...
#>  $ sex   : Factor w/ 4 levels "female","male",..: 2 1 4 3 2 1 4 3 2 1 ...
#>  $ mark  : int  72 38 8 118 22 23 1 46 35 42 ...
#>  $ catch : int  54 22 6 82 14 15 0 29 16 33 ...
#>  $ recap : int  33 7 0 40 9 5 0 14 4 5 ...
head(MuskieWI06MR)
#>     county       lake     sex mark catch recap
#> 1 Langlade Enterprise    male   72    54    33
#> 2 Langlade Enterprise  female   38    22     7
#> 3 Langlade Enterprise unknown    8     6     0
#> 4 Langlade Enterprise   total  118    82    40
#> 5   Oneida      Clear    male   22    14     9
#> 6   Oneida      Clear  female   23    15     5