Catch and effort for three Snapper species in a depletion experiment.

Catch and effort for three Snapper species (Pristipomoides zonatus, Pristipomoides auricilla, and Etelis carbunculUs) in a depletion experiment around Pathfinder Reef in the Mariana Archipelago.

Format

A data frame with 13 observations on the following 5 variables.

date

Date (1984)

effort

Fishing effort (line-hours of a bottom hand-line)

Pzonatus

Catch of Pristipomoides zonatus

Pauricilla

Catch of Pristipomoides auricilla

Ecarbunculus

Catch of Etelis carbunculUs

Source

From Table 3 of Polovina, J.J. 1985. A variable catchability version of the Leslie model with application to an intensive fishing experiment on a multispecies stock. Fishery Bulletin 84:423-428. CSV file

Topic(s)

Depletion methods

Leslie method

DeLury method

Population size

Abundance

Catchability

Examples

data(Pathfinder)
str(Pathfinder)
#> 'data.frame':	13 obs. of  5 variables:
#>  $ date        : Factor w/ 13 levels "10-Apr","11-Apr",..: 1 2 3 4 5 6 7 8 9 10 ...
#>  $ effort      : num  27.5 23.7 21.3 29.7 29.3 17.5 30.7 21.4 22.4 21.6 ...
#>  $ Pzonatus    : int  98 111 47 91 66 50 67 38 41 28 ...
#>  $ Pauricilla  : int  12 17 12 29 17 13 26 12 15 17 ...
#>  $ Ecarbunculus: int  42 22 41 19 29 21 36 15 25 15 ...
head(Pathfinder)
#>     date effort Pzonatus Pauricilla Ecarbunculus
#> 1 10-Apr   27.5       98         12           42
#> 2 11-Apr   23.7      111         17           22
#> 3 12-Apr   21.3       47         12           41
#> 4 13-Apr   29.7       91         29           19
#> 5 14-Apr   29.3       66         17           29
#> 6 15-Apr   17.5       50         13           21