Capture histories (2 samples) of Northern Pike from Harding Lake.

Capture histories for Northern Pike (Esox lucius) captured from Harding Lake, 1990.

Format

A data frame with 481 observations on the following 3 variables.

fish

a numeric vector of unique fish identification numbers

first

a numeric vector of indicator variables for the first sample (1=captured)

second

a numeric vector of indicator variables for the second sample (1=captured)

Source

Capture histories simulated from summarized data in table 2 and text of Burkholder, A. 1991. Abundance and composition of northern pike, Harding Lake, 1990. Fishery Data Series 91-9, Alaska Department of Fish and Game. [Was (is?) from http://www.sf.adfg.state.ak.us/FedAidpdfs/Fds91-09.pdf.] CSV file

Note

Only Northern Pike >449 mm were considered here.

Topic(s)

Population Size

Abundance

Mark-Recapture

Capture-Recapture

Petersen

Capture History

Examples

data(PikeHL)
str(PikeHL)
#> 'data.frame':	481 obs. of  3 variables:
#>  $ fish  : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ first : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ second: int  0 0 0 0 0 0 0 0 0 0 ...
head(PikeHL)
#>   fish first second
#> 1    1     1      0
#> 2    2     1      0
#> 3    3     1      0
#> 4    4     1      0
#> 5    5     1      0
#> 6    6     1      0