Biological data for Pygmy Whitefish from Dina Lake #1 (British Columbia), 2000 and 2001.

Biological information for Pygmy Whitefish (Prosopium coulterii) from Dina Lake #1 (British Columbia), 2000 and 2001.

Format

A data frame with 690 observations on the following 13 variables.

year

Year of capture (2000, 2001)

month

Month of capture

week

Week within a month of capture

net_no

Unique net identification number

fish_no

Unique fish identification number

fl

Fork length (cm)

tl

Total length (cm)

wt

Weight (g)

sex

Sex code (F=Female, M=Male, Imm=immature)

mat

Maturity code (Imm=immature, MG=maturing, MT=mature)

scale_age

Scale age (in years)

oto_age

Otolith age (in years)

USE

Should the individual be used for analyses.

Source

2000 data from Appendix 9 of J. D. McPhail and R. J. Zemlak. October 2001. Pygmy Whitefish studies on Dina Lake #1, 2000. Peace/Williston Fish and Wildlife Compensation Program, Report No. 245. 36pp plus appendices.

2001 data from Appendix 14 of Zemlak, R.J. and J.D. McPhail. 2003. Pygmy Whitefish studies on Dina Lake #1, 2001. Peace/Williston Fish and Wildlife Compensation Program Report No. 279. 35pp plus appendices. [Was (is?) from http://a100.gov.bc.ca/appsdata/acat/documents/r39503/pwfwcp_report_no_279_1383686631923_d6d1b25e3156cee45e442c5b3f756a6a7b4bb7ad7753efb1c51dc23ac997a9a6.pdf.] CSV file

Note

The tl and USE variables were added or teaching purposes only. The tl was created from fl using a known fl-tl relationship and a small random error.

Topic(s)

Weight-Length

Age Comparison

Condition

Length Frequency

Examples

data(PygmyWFBC)
str(PygmyWFBC)
#> 'data.frame':	690 obs. of  13 variables:
#>  $ year     : int  2001 2001 2001 2001 2001 2001 2001 2001 2001 2001 ...
#>  $ month    : Factor w/ 7 levels "August","July",..: 4 4 4 4 3 3 3 3 3 3 ...
#>  $ week     : int  4 4 4 4 1 1 1 1 1 1 ...
#>  $ net_no   : int  7 7 7 7 3 3 4 4 4 4 ...
#>  $ fish_no  : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ fl       : num  10.6 10.2 11.3 12.5 11.9 11.9 9.5 9.5 9.6 9.7 ...
#>  $ tl       : num  11.2 10.8 12.6 13.4 12.2 12.6 10.2 10.1 10.2 10.1 ...
#>  $ wt       : num  12.6 11.3 13.4 16.4 16.6 18.3 8.6 9.2 9.1 10 ...
#>  $ sex      : Factor w/ 2 levels "F","M": 1 2 1 NA 1 1 1 2 2 1 ...
#>  $ mat      : Factor w/ 3 levels "IMM","MG","MT": 2 2 3 NA 2 2 2 2 2 2 ...
#>  $ scale_age: int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ oto_age  : int  2 4 5 NA 3 3 1 1 2 1 ...
#>  $ USE      : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
head(PygmyWFBC)
#>   year month week net_no fish_no   fl   tl   wt  sex  mat scale_age oto_age
#> 1 2001   May    4      7       1 10.6 11.2 12.6    F   MG        NA       2
#> 2 2001   May    4      7       2 10.2 10.8 11.3    M   MG        NA       4
#> 3 2001   May    4      7       3 11.3 12.6 13.4    F   MT        NA       5
#> 4 2001   May    4      7       4 12.5 13.4 16.4 <NA> <NA>        NA      NA
#> 5 2001  June    1      3       5 11.9 12.2 16.6    F   MG        NA       3
#> 6 2001  June    1      3       6 11.9 12.6 18.3    F   MG        NA       3
#>    USE
#> 1 TRUE
#> 2 TRUE
#> 3 TRUE
#> 4 TRUE
#> 5 TRUE
#> 6 TRUE