Catch and effort data for Deepwater Caridean Shrimp.

Catch and effort data for Deepwater Caridean Shrimp (Heterocarpus laevigatus) from 15 days in 1984 from near Alamagan Islan in the Marian Archipelago (near Guam).

Format

A data frame with 15 observations on the following 4 variables.

day

Day of the catch. Day 9 is 9-Jan-1984.

standard

Catch (kg) of of Shrimp in the standard traps.

pyramid

Catch (kg) of of Shrimp in the pyramid traps.

effort

Total effort (trap-nights) for the standard traps.

Source

From Table 1 of Ralson, S. 1986. An intensive fishing experiment for the Caridean Shrimp, Heterocarpus laevigatus, at Alamagn Island in the Mariana Archipelago. Fishery Bulletin 84:927-934. CSV file

Details

Catch (kg) and effort (trap-nights) of Deepwater Caridean Shrimp (Heterocarpus laevigatus) from 15 days in 1984 from near Alamagan Islan in the Marian Archipelago (near Guam). The data start on 9-Jan-1984. Catches were recorded separately for standard traps and in pyramid traps.

The original authors estimated populations size using the Leslie method with the cumulative catch from the combined catch in the standard and pyramid traps, but with a CPE computed from just the catch in standard traps.

Topic(s)

Population size

Abundance

Depletion methods

Leslie method

DeLury method

Catchability

Examples

data(ShrimpGuam)
str(ShrimpGuam)
#> 'data.frame':	15 obs. of  4 variables:
#>  $ day     : int  9 10 11 12 13 14 15 16 17 18 ...
#>  $ standard: num  32.5 42.3 64.6 45 85.8 ...
#>  $ pyramid : num  0 37.65 0 23.86 6.72 ...
#>  $ effort  : int  20 18 20 17 20 19 15 14 19 20 ...
head(ShrimpGuam)
#>   day standard pyramid effort
#> 1   9    32.52    0.00     20
#> 2  10    42.30   37.65     18
#> 3  11    64.61    0.00     20
#> 4  12    44.96   23.86     17
#> 5  13    85.80    6.72     20
#> 6  14    42.90   22.98     19

## Computations by the original authors
# CPE for just the standard traps
ShrimpGuam$CPE <- ShrimpGuam$standard/ShrimpGuam$effort
# Total catch in both traps
ShrimpGuam$total <- ShrimpGuam$standard+ShrimpGuam$pyramid
# Cumulative catch in both traps (with the Ricker modification)
ShrimpGuam$cumCatch <- cumsum(ShrimpGuam$total)-ShrimpGuam$total/2