Ages (subsample) and lengths (all fish) for Spot.

As many as 10 fish per 1-inch total length intervals from the SpotVA1 data frame were obtained for age assignment. The remaining fish in the file were only measured for length (i.e., the ages were deleted). This data file can be used to demonstrate the use of age-length keys.

Format

A data frame of 403 observations on the following 2 variables:

tl

Measured total lengths (in inches)

age

Ages assigned from examination of otoliths

Source

From Table 1 in Chapter 8 (Spot) of the VMRC Final Report on Finfish Ageing, 2002 by the Center for Quantitative Fisheries Ecology at Old Dominion University.

Topic(s)

Age-Length Key

See also

Examples

data(SpotVA2)
str(SpotVA2)
#> 'data.frame':	403 obs. of  2 variables:
#>  $ tl : num  10.6 7.1 12.3 9.7 11.2 8.9 12.6 7.6 10 7 ...
#>  $ age: int  1 1 3 2 3 1 3 1 1 1 ...
head(SpotVA2)
#>     tl age
#> 1 10.6   1
#> 2  7.1   1
#> 3 12.3   3
#> 4  9.7   2
#> 5 11.2   3
#> 6  8.9   1

## Extract the aged sample
spot.aged <- subset(SpotVA2,!is.na(age))
str(spot.aged)
#> 'data.frame':	72 obs. of  2 variables:
#>  $ tl : num  10.6 7.1 12.3 9.7 11.2 8.9 12.6 7.6 10 7 ...
#>  $ age: int  1 1 3 2 3 1 3 1 1 1 ...

## Extract the length sample
spot.length <- subset(SpotVA2,is.na(age))
str(spot.length)
#> 'data.frame':	331 obs. of  2 variables:
#>  $ tl : num  9.6 9.4 9.1 9.4 9.6 9 8.2 9.8 10.7 9.1 ...
#>  $ age: int  NA NA NA NA NA NA NA NA NA NA ...