Summarized multiple mark-recapture data for Redear Sunfish.

The number of Redear Sunfish (Lepomis microlophus) caught in multiple samples from Gordy Lake, IN. The caught fish were examined for previous marks, marked (if previously unmarked), and then returned to the population.

Format

A data frame with 6 observations on the following 4 variables:

t

Sample number

caught

Total number of fish caught in the sample

recaps

Number of previously marked fish in the sample

retmarks

Number of marked fish returned to the population

Source

Originally from Gerking, S.D. 1953. Vital statistics of the fish population of Gordy Lake, Indiana. Transactions of the American Fisheries Society. 82:48-67. But also found in Table 2.4 of Krebs, C.J. 1999. Ecological Methodology. Addison-Welsey Educational Publishing, second edition and Table 4.4 of Ricker, W.E. 1975. Computation and interpretation of biological statistics of fish populations. Technical Report Bulletin 191, Bulletin of the Fisheries Research Board of Canada. CSV file

Topic(s)

Population Size

Abundance

Mark-Recapture

Capture-Recapture

Schnabel

Schumacher-Eschmeyer

Examples

data(SunfishIN)
str(SunfishIN)
#> 'data.frame':	14 obs. of  4 variables:
#>  $ t       : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ caught  : int  10 27 17 7 1 5 6 15 9 18 ...
#>  $ recaps  : int  0 0 0 0 0 0 2 1 5 5 ...
#>  $ retmarks: int  10 27 17 7 1 5 6 15 9 18 ...
SunfishIN
#>     t caught recaps retmarks
#> 1   1     10      0       10
#> 2   2     27      0       27
#> 3   3     17      0       17
#> 4   4      7      0        7
#> 5   5      1      0        1
#> 6   6      5      0        5
#> 7   7      6      2        6
#> 8   8     15      1       15
#> 9   9      9      5        9
#> 10 10     18      5       18
#> 11 11     16      4       14
#> 12 12      5      2        5
#> 13 13      7      2        6
#> 14 14     19      3        0