Biological data for Walleye from Lake Erie, 2003-2014.

Walleye (Sander vitreus) biological data (length, weight, sex, maturity, and age) from several locations in Lake Erie, October-November of 2003-2014.

Format

A data frame of 33734 observations on the following 10 variables:

setID

Unique gear (multifilament gill net kegged 6 ft below surface) set identification number.

loc

Regional location (1=Toledo to Huron, 2=Huron to Fairport Harbor, 3=Fairport Harbor to Conneaut).

grid

2.5-minute sampling grid location.

year

Year of data.

tl

Total length (mm).

w

Weight (g). There are several missing values.

sex

Sex (female, male).

mat

Maturity (immature, mature).

age

Age (yrs) from otoliths.

Source

These unpublished data are from the Ohio Department of Natural Resources, Division of Wildlife (via Christopher Vandergoot). Do not use for other than educational purposes without permission from the source.

Topic(s)

Growth

von Bertalanffy

Weight-Length

Catch curve

Mortality

Maturity

Size Structure

Length Frequency

Condition

Examples

data(WalleyeErie2)
str(WalleyeErie2)
#> 'data.frame':	33734 obs. of  9 variables:
#>  $ setID: int  2003001 2003001 2003001 2003001 2003001 2003001 2003001 2003001 2003001 2003001 ...
#>  $ loc  : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ grid : int  940 940 940 940 940 940 940 940 940 940 ...
#>  $ year : int  2003 2003 2003 2003 2003 2003 2003 2003 2003 2003 ...
#>  $ tl   : int  360 371 375 375 375 376 379 379 380 380 ...
#>  $ w    : int  460 571 507 584 537 553 500 522 525 543 ...
#>  $ sex  : Factor w/ 2 levels "female","male": 2 2 2 2 2 2 1 2 2 2 ...
#>  $ mat  : Factor w/ 2 levels "immature","mature": 2 2 2 2 2 2 1 2 2 2 ...
#>  $ age  : int  2 2 2 2 2 2 2 2 2 2 ...
head(WalleyeErie2)
#>     setID loc grid year  tl   w  sex    mat age
#> 1 2003001   1  940 2003 360 460 male mature   2
#> 2 2003001   1  940 2003 371 571 male mature   2
#> 3 2003001   1  940 2003 375 507 male mature   2
#> 4 2003001   1  940 2003 375 584 male mature   2
#> 5 2003001   1  940 2003 375 537 male mature   2
#> 6 2003001   1  940 2003 376 553 male mature   2
xtabs(~year+loc+sex,data=WalleyeErie2)
#> , , sex = female
#> 
#>       loc
#> year      1    2    3
#>   2003  106  169   12
#>   2004  286  367  147
#>   2005  338  820   98
#>   2006  246 1156  343
#>   2007  315  349   65
#>   2008  370  746  191
#>   2009  296  838  270
#>   2010  326  921  175
#>   2011  268  735  398
#>   2012  502  629  169
#>   2013  235  778   94
#>   2014  325  207    0
#> 
#> , , sex = male
#> 
#>       loc
#> year      1    2    3
#>   2003  268  255   19
#>   2004  345  555  139
#>   2005  468 1488  123
#>   2006  448 1705  162
#>   2007  740  967   91
#>   2008  578 1396  168
#>   2009  751 1928  167
#>   2010  564 1320  127
#>   2011  585 1112  324
#>   2012  917  882  148
#>   2013  317  776  137
#>   2014  289  185    0
#>