Summarized multiple mark-recapture data for Walleye from four lakes in Northern Minnesota.

Summary results of capture histories (number captured, number of recaptured fish, and number of unmarked fish that were marked) for Walleye (Sander vitreus) collected from four lakes in Northern Minnesota, USA.

Format

A data frame with 20 observations on the following 5 variables.

lake

Studied lake (Crooked, Fourmile, Island, Tom)

date

Capture date

catch

Total fish captured in each sample

recap

Marked fish captured in each sample

retMark

Marked fish returned to the population

Source

From appendix 2 in Borkholder, B.D., A.J. Edwards, and C. Olson. 2007. Spring adult and fall juvenile walleye popluation surveys within the 1854 ceded territory of Minnesota, 2006. Fond du Lac Division of Resource Management, Technical Report 41. CSV file

Topic(s)

Population Size

Abundance

Mark-Recapture

Capture-Recapture

Schnabel

Schumacher-Eschmeyer

Examples

data(WalleyeMN06b)
str(WalleyeMN06b)
#> 'data.frame':	20 obs. of  5 variables:
#>  $ lake   : Factor w/ 4 levels "Crooked","Fourmile",..: 3 3 3 3 3 3 3 3 3 1 ...
#>  $ date   : Factor w/ 12 levels "17-Apr","18-Apr",..: 1 2 3 4 5 6 7 8 9 5 ...
#>  $ catch  : int  597 743 813 697 476 521 571 465 116 145 ...
#>  $ recap  : int  NA 52 125 186 213 164 77 53 2 NA ...
#>  $ retMark: int  597 690 688 511 253 347 484 402 NA 145 ...
head(WalleyeMN06b)
#>     lake   date catch recap retMark
#> 1 Island 17-Apr   597    NA     597
#> 2 Island 18-Apr   743    52     690
#> 3 Island 19-Apr   813   125     688
#> 4 Island 20-Apr   697   186     511
#> 5 Island 21-Apr   476   213     253
#> 6 Island 22-Apr   521   164     347