Summarized multiple mark-recapture data for YOY walleye.

The numbers of young-of-year walleye (Sander vitreus) that were captured, found to have previous marks (i.e., recaptured), and were newly marked on several sampling occasions in 1959, 1960, 1961, and 1962 in Pike Lake, Wisconsin.

Format

A data frame with 33 observations on the following 5 variables:

year

Sampling year

t

Sampling occasion within each year

caught

Number of walleye captured

recaptures

Number of marked walleyes captured

newmarks

Number of unmarked walleyes that were captured, marked, and returned to the population

Source

From table 3 of Mraz, D. 1968. Recruitment, growth, exploitation, and management of walleyes in a southeastern Wisconsin lake. Wisconsin Department of Natural Resources Technical Bulletin 40. 38 pages. CSV file

Topic(s)

Population Size

Abundance

Mark-Recapture

Capture-Recapture

Schnabel

Schumacher-Eschmeyer

Examples

data(WalleyePL)
str(WalleyePL)
#> 'data.frame':	33 obs. of  5 variables:
#>  $ year      : int  1959 1959 1959 1959 1959 1959 1959 1959 1959 1959 ...
#>  $ t         : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ caught    : int  145 99 145 155 177 38 308 179 290 179 ...
#>  $ recaptures: int  0 15 9 10 40 4 64 41 68 54 ...
#>  $ newmarks  : int  145 84 136 145 137 34 244 138 222 125 ...
WalleyePL
#>    year  t caught recaptures newmarks
#> 1  1959  1    145          0      145
#> 2  1959  2     99         15       84
#> 3  1959  3    145          9      136
#> 4  1959  4    155         10      145
#> 5  1959  5    177         40      137
#> 6  1959  6     38          4       34
#> 7  1959  7    308         64      244
#> 8  1959  8    179         41      138
#> 9  1959  9    290         68      222
#> 10 1959 10    179         54      125
#> 11 1959 11    248         68        0
#> 12 1960  1    141          0      141
#> 13 1960  2    110          6      104
#> 14 1960  3     88          8       80
#> 15 1960  4     97         13       84
#> 16 1960  5     55         11       44
#> 17 1960  6    105         32       73
#> 18 1960  7     19          7       12
#> 19 1960  8    163         44      119
#> 20 1960  9    143         47       96
#> 21 1960 10    106         30       76
#> 22 1961  1    352          0      352
#> 23 1961  2    291         31      260
#> 24 1961  3    548         87      461
#> 25 1961  4    372         96      276
#> 26 1961  5    380         95      285
#> 27 1961  6    260        109      151
#> 28 1961  7    335        135        0
#> 29 1962  1    321          0      321
#> 30 1962  2    412         45      367
#> 31 1962  3    178         55      123
#> 32 1962  4    415         93      322
#> 33 1962  5    367        113        0
subset(WalleyePL,year==1960)
#>    year  t caught recaptures newmarks
#> 12 1960  1    141          0      141
#> 13 1960  2    110          6      104
#> 14 1960  3     88          8       80
#> 15 1960  4     97         13       84
#> 16 1960  5     55         11       44
#> 17 1960  6    105         32       73
#> 18 1960  7     19          7       12
#> 19 1960  8    163         44      119
#> 20 1960  9    143         47       96
#> 21 1960 10    106         30       76