Landings and value of Lake Superior Lake Whitefish.

Landings and value of Lake Whitefish Coregonus clupeaformis in Lake Superior by state from 1971-2001.

Format

A data frame with 12 observations on the following 4 variables:

year

Year of catch.

state

State of catch (MI, MN, or WI).

catch

Catch in lbs.

value

Value of catch in dollars.

Source

Results from a query in 2006 to the NMFS Great Lakes Commercial Fishery Landings database which was (is?) at http://www.st.nmfs.noaa.gov/pls/webpls/webst1.MF_GL_SPECIES_HELP.SPECIES.

Topic(s)

Other

Examples

data(WhitefishLS)
str(WhitefishLS)
#> 'data.frame':	73 obs. of  4 variables:
#>  $ year : int  1971 1971 1972 1972 1973 1973 1973 1974 1974 1974 ...
#>  $ state: Factor w/ 3 levels "MI","MN","WI": 1 3 1 3 1 2 3 1 2 3 ...
#>  $ catch: int  447564 264009 455631 277295 501436 3064 260482 499723 2113 264575 ...
#>  $ value: int  229001 110540 250860 133918 301909 1568 137064 340225 1070 140247 ...
head(WhitefishLS)
#>   year state  catch  value
#> 1 1971    MI 447564 229001
#> 2 1971    WI 264009 110540
#> 3 1972    MI 455631 250860
#> 4 1972    WI 277295 133918
#> 5 1973    MI 501436 301909
#> 6 1973    MN   3064   1568